Remove stale remote control size exception
